refactor: consolidate config into core
- Move Config/Manager to dare_framework.core.config
- Move ComponentType to dare_framework.contracts
- Update all imports and plugin managers to use new core config
- Remove legacy ConfigProvider/LayeredConfigProvider components
- Archive consolidate-config-to-core change
