--- dare_framework/components/config_providers/__init__.py
@@ -1,1 +1,0 @@
-"""Config provider implementations loaded via the `dare_framework.v2.config_providers` entrypoint group."""
