+++ dare_framework/components/hooks/noop.py
@@ -4,7 +4,7 @@ from typing import Any
 
 from dare_framework.components.base_component import ConfigurableComponent
 from dare_framework.core.hook.models import HookPhase
-from dare_framework.components.plugin_system.component_type import ComponentType
+from dare_framework.contracts import ComponentType
 
 from .protocols import IHook
