+++ dare_framework/components/model_adapters/mock.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Iterable
 
 from dare_framework.contracts.model import IModelAdapter, Message, ModelResponse
-from dare_framework.components.plugin_system.component_type import ComponentType
+from dare_framework.contracts import ComponentType
 from ..base_component import ConfigurableComponent
