+++ dare_framework/components/remediators/noop.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dare_framework.components.base_component import ConfigurableComponent
 from dare_framework.core.protocols import IRemediator
 from dare_framework.core.plan.results import VerifyResult
-from dare_framework.components.plugin_system.component_type import ComponentType
+from dare_framework.contracts import ComponentType
 
 
 class NoOpRemediator(ConfigurableComponent, IRemediator):
