+++ dare_framework/components/skills/noop.py
@@ -5,7 +5,7 @@ from typing import Any
 from dare_framework.components.base_component import ConfigurableComponent
 from dare_framework.contracts.run_context import RunContext
 from dare_framework.contracts.tool import ToolResult
-from dare_framework.components.plugin_system.component_type import ComponentType
+from dare_framework.contracts import ComponentType
 
 from .protocols import ISkill
