+++ dare_framework/components/tools/run_command.py
@@ -9,7 +9,7 @@ from dare_framework.contracts.ids import generator_id
 from dare_framework.contracts.risk import RiskLevel
 from dare_framework.contracts.run_context import RunContext
 from dare_framework.contracts.tool import ToolResult, ToolType
-from dare_framework.components.plugin_system.component_type import ComponentType
+from dare_framework.contracts import ComponentType
 from ..base_component import ConfigurableComponent
