--- dare_framework/config/__init__.py
@@ -1,8 +1,0 @@
-"""Configuration surface (v2)."""
-
-from .config import Config
-from .component_config import ComponentConfig
-from .llm_config import LLMConfig
-
-__all__ = ["Config", "ComponentConfig", "LLMConfig"]
-
