--- dare_framework/config/component_config.py
@@ -1,26 +1,0 @@
-from __future__ import annotations
-
-from dataclasses import dataclass, field
-from typing import Any
-
-
-@dataclass(frozen=True)
-class ComponentConfig:
-    """Per-component-type configuration with a disabled list and named entries."""
-
-    disabled: list[str] = field(default_factory=list)
-    entries: dict[str, Any] = field(default_factory=dict)
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> "ComponentConfig":
-        disabled_raw = data.get("disabled", [])
-        disabled = [str(item) for item in disabled_raw] if isinstance(disabled_raw, list) else []
-        entries = {key: value for key, value in data.items() if key != "disabled"}
-        return cls(disabled=disabled, entries=entries)
-
-    def to_dict(self) -> dict[str, Any]:
-        payload = dict(self.entries)
-        if self.disabled:
-            payload["disabled"] = list(self.disabled)
-        return payload
-
