+++ dare_framework/contracts/__init__.py
@@ -5,5 +5,6 @@ from .model import IModelAdapter, ModelResponse
 from .run_context import RunContext
 from .risk import RiskLevel
 from .evidence import Evidence
+from .component_type import ComponentType
 
-__all__ = ["ITool", "IModelAdapter", "ModelResponse", "RunContext", "ToolResult", "RiskLevel", "Evidence"]
+__all__ = ["ITool", "IModelAdapter", "ModelResponse", "RunContext", "ToolResult", "RiskLevel", "Evidence", "ComponentType"]
