+++ dare_framework/core/config/__init__.py
@@ -1,0 +1,13 @@
+"""Core configuration management (v2)."""
+
+from .models import Config, LLMConfig, ComponentConfig
+from .manager import ConfigManager, merge_config_layers, build_config_from_layers
+
+__all__ = [
+    "Config",
+    "LLMConfig",
+    "ComponentConfig",
+    "ConfigManager", 
+    "merge_config_layers",
+    "build_config_from_layers",
+]
