+++ tests/unit/test_layered_config_provider.py
@@ -1,16 +1,16 @@
-from dare_framework.components.config_providers.layered_config_provider import LayeredConfigProvider
+from dare_framework.core.config.manager import ConfigManager
 
 
 def test_layered_config_merge_precedence():
-    provider = LayeredConfigProvider(
+    manager = ConfigManager(
         system={"llm": {"model": "sys", "temperature": 0.1}},
         project={"llm": {"model": "project"}},
         user={"llm": {"model": "user"}},
         session={"runtime": {"timeout": 10}},
     )
 
-    assert provider.get("llm.model") == "user"
-    assert provider.get("llm.temperature") == 0.1
-    assert provider.get_namespace("runtime") == {"timeout": 10}
-    assert provider.config_hash
-    assert "session" in provider.sources
+    assert manager.get("llm.model") == "user"
+    assert manager.get("llm.temperature") == 0.1
+    assert manager.get_namespace("runtime") == {"timeout": 10}
+    assert manager.config_hash
+    assert "session" in manager.sources
